refactor: improve coefficient_of_variation calculation for type safety
Updated the return type of the coefficient_of_variation calculation in the ActivityAggregator service to ensure it consistently returns a float. This change enhances type safety and aligns with the goal of maintaining a robust and maintainable codebase for the psychiatry platform.

+++ src/big_mood_detector/domain/services/activity_aggregator.py
@@ -261,7 +261,7 @@ class ActivityAggregator:
         coefficient_of_variation = std_dev / mean_intensity
 
         # Scale to 0-1 range (CV > 1.5 maps to 1)
-        return min(coefficient_of_variation / 1.5, 1.0)
+        return float(min(coefficient_of_variation / 1.5, 1.0))
 
     def _calculate_activity_hours(
         self, steps_records: list[ActivityRecord]
